started working on vacancies, in progress
