+++ bewelder_redesign/settings.py
@@ -25,7 +25,7 @@ SECRET_KEY = 'vd0l%=zj66k$^ob5zdwwvxmfiiop198scrjbh*16*2^y(h=xra'
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = bool(os.environ.get('DJANGO_DEBUG', True))
 
-ALLOWED_HOSTS = ['bewgb.ru', 'bewgb.local']
+ALLOWED_HOSTS = ['bewgb.ru', 'bewgb.local', '127.0.0.1']
 
 
 # Application definition
