+++ mainapp/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
-from .models import Vacancy, Resume, Category
+# from .models import Vacancy, Resume, Category
 # Register your models here.
-admin.site.register(Category)
-admin.site.register(Vacancy)
-admin.site.register(Resume)
+# admin.site.register(Category)
+# admin.site.register(Vacancy)
+# admin.site.register(Resume)
